refactor(T9357): remove resolveAnthropicApiKey shim from adapters — D-ph4-02 final close
Delete the local resolveAnthropicApiKey() function from packages/adapters/src/providers/claude-sdk/spawn.ts and migrate all 3 call-sites (function definition + 2 usage sites) to resolveCredentials('anthropic').apiKey. The canonical resolver is provided via a new shared module at packages/adapters/src/providers/shared/credentials.ts which implements the same 3-tier resolution chain (env → stored key → OAuth) using @cleocode/contracts parseClaudeCodeCredentials helper plus Node.js builtins. This avoids the circular dependency that would arise from importing @cleocode/core (which depends on @cleocode/adapters). All 19 claude-sdk spawn tests pass. Acceptance grep returns zero hits.
